pyproject.toml
[project]
name = "alpaca-hft-paper"
version = "0.1.0"
description = "Paper forward-testing harness for HFT-style strategies on Alpaca"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "Codex"}]

dependencies = [
  "alpaca-py==0.20.2",
  "numpy==1.26.4",
  "scikit-learn==1.5.1",
  "pytz==2024.1",
]

[project.optional-dependencies]
uvloop = ["uvloop==0.20.0"]

[project.scripts]
alpaca-hft = "src.main:main"
